+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to this project will be documented in this file.
 
+## [4.10.6]
+
 ## [4.10.5]
 
 ## [4.10.4]